refactor: theme 관련 속성 분리 및 typography 구조 추가
