+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from '@/App';
 import GlobalStyles from '@/styles/GlobalStyles';
 import { ThemeProvider } from '@emotion/react';
-import theme from '@/styles/theme';
+import theme from '@/styles/theme/index';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
