+++ src/styles/theme/fontSizes.ts
@@ -1,0 +1,20 @@
+const fontSizes = {
+  title1Bold: '20px',
+  title1Regular: '20px',
+  title2Bold: '16px',
+  title2Regular: '16px',
+  subtitle1Bold: '16px',
+  subtitle1Regular: '16px',
+  subtitle2Bold: '14px',
+  subtitle2Regular: '14px',
+  body1Bold: '16px',
+  body1Regular: '16px',
+  body2Bold: '14px',
+  body2Regular: '14px',
+  label1Bold: '14px',
+  label1Regular: '14px',
+  label2Bold: '12px',
+  label2Regular: '12px',
+};
+
+export default fontSizes;
