+++ src/styles/theme/index.ts
@@ -1,0 +1,13 @@
+import colors from '@styles/theme/colors';
+import fontSizes from '@styles/theme/fontSizes';
+import spacing from '@styles/theme/spacing';
+import typography from '@styles/theme/typography';
+
+const theme = {
+  colors,
+  fontSizes,
+  typography,
+  spacing,
+};
+
+export default theme;
