+++ src/styles/theme/typography.ts
@@ -1,0 +1,15 @@
+const typography = {
+    body1Bold: {
+      fontSize: '16px',
+      fontWeight: 700,
+      lineHeight: '24px',
+    },
+    label1Regular: {
+      fontSize: '14px',
+      fontWeight: 400,
+      lineHeight: '20px',
+    },
+  };
+  
+  export default typography;
+
