+++ src/types/emotion.d.ts
@@ -1,5 +1,8 @@
-/** @jsxImportSource @emotion/react */
 import '@emotion/react';
+import type colors from '@styles/theme/colors';
+import type fontSizes from '@styles/theme/fontSizes';
+import type spacing from '@styles/theme/spacing';
+import type typography from '@styles/theme/typography';
 
 declare module 'react' {
   interface Attributes {
@@ -7,10 +10,16 @@ declare module 'react' {
   }
 }
 
+type Colors = typeof colors;
+type FontSizes = typeof fontSizes;
+type Spacing = typeof spacing;
+type Typography = typeof typography;
+
 declare module '@emotion/react' {
   export interface Theme {
-    colors: typeof import('@/styles/theme').default.colors;
-    fontSizes: typeof import('@/styles/theme').default.fontSizes;
-    spacing: typeof import('@/styles/theme').default.spacing;
+    colors: Colors;
+    fontSizes: FontSizes;
+    spacing: Spacing;
+    typography: Typography; 
   }
 }
